# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: FileBrowser API
paths:
  /v1/files/list:
    get:
      summary: List files
      description: List files in a directory
      parameters:
        - name: volume
          in: query
          description: Volume id. If not provided, the default home volume will be used.
          required: false
          schema:
            type: integer
        - name: path
          in: query
          description: Path to the directory
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          description: Group files by a specific field
          required: false
          schema:
            $ref: '#/components/schemas/FileGroupBy'
        - name: sortBy
          in: query
          description: Sort files by a specific field
          required: false
          schema:
            $ref: '#/components/schemas/FileSortBy'
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            $ref: '#/components/schemas/SortOrder'
      responses:
        '200':
          description: A list of files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileGroup'
        default:
          description: An error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/admin/volumes:
    get:
      summary: List volumes
      description: List all volumes
      responses:
        '200':
          description: A list of volumes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
        default:
          description: An error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    FileGroupBy:
      type: string
      enum:
        - kind
        - modified
        - size
    FileSortBy:
      type: string
      enum:
        - name
        - size
        - modified
    SortOrder:
      type: string
      enum:
        - asc
        - desc
    FileGroupCollection:
      type: object
      properties:
        permissions:
          $ref: '#/components/schemas/FilePermissions'
        fileGroups:
          type: array
          items:
            $ref: '#/components/schemas/FileGroup'
    FileGroup:
      type: object
      properties:
        name:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
    File:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        modified:
          type: string
        is_dir:
          type: boolean
    FilePermissions:
      type: object
      properties:
        read:
          type: boolean
        write:
          type: boolean
        delete:
          type: boolean
        share:
          type: boolean
    Volume:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
        backend:
          $ref: '#/components/schemas/Backend'
    Backend:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          $ref: '#/components/schemas/BackendType'
    BackendType:
      type: string
      enum:
        - local
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message